+++ Code/Sandbox/Plugins/CryDesigner/Tools/Shape/DiscTool.cpp
@@ -90,12 +90,12 @@ bool DiscTool::SnapAngleBy90Degrees(float angle, float& outSnappedAngle)
 
 	if (std::abs(angle) < range)
 		outSnappedAngle = 0;
-	else if (std::abs(angle - 0.25f * PI2) < range)
-		outSnappedAngle = 0.25f * PI2;
-	else if (std::abs(angle + 0.25f * PI2) < range)
-		outSnappedAngle = -0.25f * PI2;
-	else if (std::abs(angle + 0.5f * PI2) < range)
-		outSnappedAngle = -0.5f * PI2;
+	else if (std::abs(angle - static_cast<float>(0.25 * PI2)) < range)
+		outSnappedAngle = static_cast<float>(0.25 * PI2);
+	else if (std::abs(angle + static_cast<float>(0.25 * PI2)) < range)
+		outSnappedAngle = static_cast<float>(-0.25 * PI2);
+	else if (std::abs(angle + static_cast<float>(0.5 * PI2)) < range)
+		outSnappedAngle = static_cast<float>(-0.5 * PI2);
 	else
 	{
 		outSnappedAngle = angle;
